disable add to favourite button when showing favourites
