--- db.json
@@ -1,32 +1,0 @@
-{
-	"user" : {
-		"7t7t78t88bjbbjkknk" : {
-			"list" : {
-				"favourite" : [
-					{
-						"id" : 12345678907,
-						"title" : "testing my app"
-					}
-				],
-				"shared" : [
-					{
-						"id" : 12345678907,
-						"title" : "testing my app"
-					}
-				]
-			},
-			"name" : "Okoro Celestine"
-		}
-	},
-	"article" : {
-		"12345678907" : {
-			"author": "Abhimanyu Ghoshal",
-			"title": "Facebook launches Express Wi-Fi in India to bring rural areas online",
-			"description": "After a failed effort to offer free internet access (with strings attached) to people in India, Facebook has now launched Express Wi-Fi, a service that lets users log on to Wi-Fi networks ...",
-			"url": "https://thenextweb.com/facebook/2017/05/04/facebook-launches-express-wi-fi-in-india-to-bring-rural-areas-online/",
-			"urlToImage": "https://cdn2.tnwcdn.com/wp-content/blogs.dir/1/files/2017/05/Facebook-Express-Wi-Fi.jpg",
-			"publishedAt": "2017-05-04T13:18:36Z",
-			"scrapeDetails":"After a failed effort to offer free internet access (with strings attached) to people in India, Facebook has now launched Express Wi-Fi, a service that lets users log on to Wi-Fi networks "
-		}
-	}
-}
