--- src/.env
@@ -1,6 +1,0 @@
-APIKEY=AIzaSyBRVO8XJcSp9EaxRRlsH5DKMILYygEUL3Y
-AUTH_DOMAIN=completenews-3a553.firebaseapp.com
-DATABASE_URL=https://completenews-3a553.firebaseio.com
-PROJECTID=completenews-3a553
-STORAGEBUCKET=completenews-3a553.appspot.com
-MESSAGINGSENDERID=21466984743
