+++ src/js/store/SourceStore.js
@@ -23,7 +23,6 @@ class Sources extends EventEmitter {
         this.emit('change');
         break;
       default :
-        return ('error');
     }
   }
 }
