+++ src/js/utilities/api.js
@@ -14,7 +14,7 @@ export const getSources = () => {
   });
 };
 
-export const getHeadlines = (source, sort) => {
+export const getHeadlines = ((source, sort) => {
   let apiURl = '';
   if (sort === '') {
     apiURl = `https://newsapi.org/v1/articles?source=${source
@@ -32,10 +32,10 @@ export const getHeadlines = (source, sort) => {
       reject(`Error occurred, ${error}`);
     });
   });
-};
+});
 
 export const getDbHeadlines = ((email) => {
-  console.log('APIKEY',process.env.APIKEY);
+  // console.log('APIKEY',process.env.APIKEY);
   return new Promise((resolve, reject) => {
     firebase.database().ref(`/user/${email}/favourite`).once('value')
     .then((snapshot) => {
